requirements.txt
asgiref==3.2.10
certifi==2020.6.20
dj-database-url==0.5.0
Django==3.0.8
django-heroku==0.3.1
gunicorn==20.0.4
psycopg2==2.8.5
pytz==2020.1
PyYAML==5.3.1
sqlparse==0.3.1
whitenoise==5.1.0
wincertstore==0.2
requests ==2.24.0
beautifulsoup4==4.9.1
Pillow==5.2.0
google-api-core==1.23.0
google-auth==1.23.0
google-cloud-vision==2.0.0
googleapis-common-protos==1.52.0
grpcio==1.31.0
pandas==1.1.4
django-boost==1.7.2
